+++ CHANGELOG.md
@@ -6,6 +6,11 @@ _13_07_2020_
 _13_07_2020_
 * FIX - Re-added session check due to integrator's bad implementation of SplitPaymentProcessor
 
+## VERSION 4.51.0
+_08_07_2020_
+* FEATURE - Added tracking behaviour to customize tracking from application
+* FIX - fixed CongratsDeepLink track path
+
 ## VERSION 4.50.2
 _01_07_2020_
 * FIX - No token in remedies after guessing card flow